pyproject.toml
[tool.poetry]
name = "backendsnaps"
version = "0.1.0"
description = ""
authors = ["Asger Hautop Drewsen <asgerdrewsen@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.9"
django-extensions = "^2.1"
ipython = "^7.4"
werkzeug = "^0.15.2"
django-cors-headers = "^2.5"
python-dotenv = "^0.10.1"
gunicorn = "^19.9"
psycopg2 = "^2.8"
whitenoise = "^4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
